Address PR feedbacks.
